adding cascade

+++ src/main/java/com/gabriel/pive/animals/services/BullService.java
@@ -43,6 +43,9 @@ public class BullService {
     }
 
     public void delete(Long id){
+        if (bullRepository.findById(id).isEmpty()){
+            throw new BullNotFoundException();
+        }
         bullRepository.deleteById(id);
     }
